1) Added exceptions 2) Added diagnose deleting

+++ src/main/java/com/yegorf/controllers/repos/DiagnoseRepo.java
@@ -3,5 +3,5 @@ package com.yegorf.controllers.repos;
 import com.yegorf.controllers.enities.Diagnose;
 import org.springframework.data.repository.CrudRepository;
 
-public interface DiagnoseRepo extends CrudRepository<Diagnose, Long> {
+public interface DiagnoseRepo extends CrudRepository<Diagnose, Integer> {
 }
